RatioDAO 인터페이스 추가, RatioMapper NAMESPACE 변경
